Internal cleanup: consolidate and simplify tests.
PiperOrigin-RevId: 321368769
