+++ tensorflow/python/autograph/impl/BUILD
@@ -40,30 +40,14 @@ py_library(
 tf_py_test(
     name = "api_test",
     srcs = ["api_test.py"],
+    python_version = "PY3",
+    srcs_version = "PY3",
     deps = [
         ":impl",
         "//tensorflow/python:client_testlib",
         "//tensorflow/python/autograph/core:test_lib",
         "//tensorflow/python/autograph/utils",
         "//third_party/py/numpy",
-    ],
-)
-
-py_test(
-    name = "api_py3_test",
-    srcs = ["api_py3_test.py"],
-    python_version = "PY3",
-    srcs_version = "PY3",
-    tags = [
-        "no_oss_py2",
-        "no_pip",
-        "nopip",
-    ],
-    deps = [
-        ":impl",
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python:constant_op",
-        "//tensorflow/python/autograph/core",
     ],
 )
